Update installation instructions in README.md
The installation instructions in README.md did not work on windows
systems - there were whitespace differences in the code that caused
the git apply command to not work.

The fix was done by changing the git command to include whitespace
option as fix, which causes the installation to work on windows systems
and doesn't affect the installation in Posix systems.